# Conflicts:
#	Ardoise/app/src/main/AndroidManifest.xml
#	Ardoise/app/src/main/java/com/example/ardoise/model/Evenement.java

+++ Ardoise/app/src/main/java/com/example/ardoise/model/Depense.java
@@ -45,7 +45,7 @@ public class Depense implements Parcelable{
         this.titre = titre;
     }
 
-    public double getMontant() {
+    public int getMontant() {
         return montant;
     }
